requirements.txt
langchain
langchain_core
langchain-openai
python-dotenv
ipykernel
langgraph
typing
